Hybrid/requirements.txt
azure-mgmt-resource~=2.0.0rc1
azure-mgmt-network~=2.0.0rc2
azure-mgmt-compute~=4.0.0rc2
haikunator
